+++ RH_RF24.h
@@ -924,7 +924,7 @@ public:
     /// \param[in] buf Location to copy the received message
     /// \param[in,out] len Pointer to available space in buf. Set to the actual number of octets copied.
     /// \return true if a valid message was copied to buf
-    bool        recv(uint8_t* buf, uint8_t* len);
+    bool        recv(uint8_t* buf, uint16_t* len);
 
     /// Waits until any previous transmit packet is finished being transmitted with waitPacketSent().
     /// Then loads a message into the transmitter and starts the transmitter. Note that a message length
