+++ examples/raspi/rf95/rf95_reliable_datagram_server/rf95_reliable_datagram_server.cpp
@@ -104,7 +104,7 @@ int main (int argc, const char* argv[] )
     {
       // Wait for a message addressed to us from the client
       uint8_t len = sizeof(buf);
-      uint8_t from;
+      uint32_t from;
       if (manager.recvfromAck(buf, &len, &from))
       {
 #ifdef RFM95_LED
