+++ examples/raspi/rf95/rf95_router_client/rf95_router_client.cpp
@@ -122,7 +122,7 @@ int main (int argc, const char* argv[] )
         // It has been reliably delivered to the next node.
         // Now wait for a reply from the ultimate server
         uint8_t len = sizeof(buf);
-        uint8_t from;
+        uint32_t from;
         if (manager.recvfromAckTimeout(buf, &len, 3000, &from))
         {
             Serial.print("got reply from : 0x");
